Removed WordAddIn Folder
The WordAddIn folder didn't work when transferred over GitHub anyway.  For testing with the Word Add-in, please download it from https://drive.google.com/file/d/1TdcpesIlVSHTgWKhlvEgFwMQ6GNw20M0/view?usp=sharing

Then, un-zip the folder and place the .jar file inside (so that the Jar file is in the same directory as "My Office Add-in").